class refactor

+++ src/features/classe/presentation/components/ClassesTable.tsx
@@ -14,9 +14,14 @@ import Link from "next/link";
 import { Button } from "@/core/components/ui/button";
 import { classeRepository } from "@/features/classe/application/repository/classe-repository";
 import { ClassType } from "../../domain/class-schema";
-import { Pen } from "lucide-react";
+import { Delete, Pen } from "lucide-react";
+import { VisibilityType } from "../../domain/visibility-schema";
+import VisibilitySwitch from "@/features/cours-sequence/presentation/components/VisibilitySwitch";
 
-export default function ClassesTable(props: { classes: ClassType[] }) {
+export default function ClassesTable(props: {
+  classes: ClassType[];
+  userId: string;
+}) {
   const { setClasseId } = classeRepository.useDeleteClasse();
   const handleDelete = async (id: string) => {
     setClasseId(id);
@@ -30,10 +35,12 @@ export default function ClassesTable(props: { classes: ClassType[] }) {
             <TableHead>Name</TableHead>
             <TableHead>Description</TableHead>
             <TableHead>Actions</TableHead>
+            <TableHead>Show</TableHead>
           </TableRow>
         </TableHeader>
         <TableBody>
           {props.classes.map((classe) => {
+          
             return (
               <TableRow key={classe.id} className="cursor-pointer ">
                 <Link href={`/classes/class/${classe.id}`} legacyBehavior>
@@ -47,26 +54,33 @@ export default function ClassesTable(props: { classes: ClassType[] }) {
                   <TableCell>{classe.description}</TableCell>
                 </Link>
 
-                <TableCell className="flex justify-center">
+                <TableCell className="flex justify-center items-center gap-4">
                   {/* Delete */}
-                  <Button
-                    data-testid="delete-class"
-                    className="mr-2"
+                  <Delete
+                    className="cursor-pointer text-red-500"
                     onClick={async () => {
-                      await handleDelete(classe.id?.toString() || "");
+                      const confirmed = confirm(
+                        "Are you sure you want to delete this class?"
+                      );
+                      if (confirmed) {
+                        await handleDelete(classe.id?.toString() || "");
+                      }
                     }}
-                    variant="destructive"
-                  >
-                    -
-                  </Button>
+                    size={20}
+                  />
 
                   {/* Edit */}
 
                   <Link href={`/classes/class/${classe.id}`}>
-                    <Button data-testid="edit-class" variant="link">
-                      <Pen size={16} />
-                    </Button>
+                    <Pen size={16} />
                   </Link>
+                </TableCell>
+                <TableCell>
+                  <VisibilitySwitch
+                    userId={props.userId}
+                    type="classe"
+                    typeId={classe.id}
+                  />
                 </TableCell>
               </TableRow>
             );
